Issue #214 - Prevent segfault with large vectors
Large vectors are allocated directly on the heap but may contain elements that are still on the stack. To ensure all of these elements are transported to the heap during the next minor GC, a special Opaque object is placed into the mutation table to ensure the whole vector is scanned for stack objects during the next minor GC.

Previously this was not done so objects from the "old" stack were still in memory and any changes to them were trashing random portions of the stack!